+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 - Week Number action to Date Time tool.
 - Fluent theme, other windows (settings, management, etc) and menus now match users Windows theme.
+- Pinned by default setting.
 - 
 
 ### Changed